requirements.txt
Kivy
numpy
trimesh
spatialmath-python 
Flask
open3d
opencv-python
pyrender
rhovee
graphslam
g2o-python
torch
torchvision